@@ -20,16 +20,16 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'gerenciador-estudos'`, () => {
+  it(`should have as title 'study-manager'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('gerenciador-estudos');
+    expect(app.title).toEqual('study-manager');
   });
 
   it('should render title', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.nativeElement as HTMLElement;
-    expect(compiled.querySelector('h1')?.textContent).toContain('Hello, gerenciador-estudos');
+    expect(compiled.querySelector('h1')?.textContent).toContain('Hello, study-manager');
   });
 });
