@@ -1,38 +1,38 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { MaterialModule } from './modules/material/material.module';
+import { LoginComponent } from './components/autenticacao/login/login.component';
+import { CadastroComponent } from './components/autenticacao/cadastro/cadastro.component';
+import { TelaInicialComponent } from './components/autenticacao/tela-inicial/tela-inicial.component';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
-import { UsuarioDadosComponent } from './components/usuario-dados/usuario-dados.component';
-import { UsuarioDisciplinasComponent } from './components/usuario-disciplinas/usuario-disciplinas.component';
-import { UsuarioProgressoComponent } from './components/usuario-progresso/usuario-progresso.component';
-import { UsuarioTarefasComponent } from './components/usuario-tarefas/usuario-tarefas.component';
-import { TelaInicialComponent } from './components/tela-inicial/tela-inicial.component';
-import { TelaInicialUsuarioComponent } from './components/tela-inicial-usuario/tela-inicial-usuario.component';
-import { LoginComponent } from './components/login/login.component';
-import { CadastroComponent } from './components/cadastro/cadastro.component';
-
+import { MaterialModule } from './modules/material/material.module';
+import { TelaUsuarioComponent } from './components/usuario/tela-usuario/tela-usuario.component';
+import { DadosComponent } from './components/usuario/dados/dados.component';
+import { DisciplinasComponent } from './components/usuario/disciplinas/disciplinas.component';
+import { TarefasComponent } from './components/usuario/tarefas/tarefas.component';
+import { ProgressoComponent } from './components/usuario/progresso/progresso.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    UsuarioDadosComponent,
-    UsuarioDisciplinasComponent,
-    UsuarioProgressoComponent,
-    UsuarioTarefasComponent,
-    TelaInicialComponent,
-    TelaInicialUsuarioComponent,
     LoginComponent,
-    CadastroComponent
+    CadastroComponent,
+    TelaInicialComponent,
+    TelaUsuarioComponent,
+    DadosComponent,
+    DisciplinasComponent,
+    TarefasComponent,
+    ProgressoComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    MaterialModule,
+    MaterialModule
   ],
   providers: [
-    provideAnimationsAsync(),
+    provideAnimationsAsync()
   ],
   bootstrap: [AppComponent]
 })
