@@ -1,16 +1,15 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { Usuario } from '../../shared/models/usuario';
-import { LocalStorageService } from '../../shared/services/local-storage.service';
-import { UsuarioService } from '../../shared/services/usuario.service';
-
+import { Usuario } from '../../../shared/models/usuario';
+import { LocalStorageService } from '../../../shared/services/local-storage.service';
+import { UsuarioService } from '../../../shared/services/usuario.service';
 
 @Component({
-  selector: 'app-usuario-dados',
-  templateUrl: './usuario-dados.component.html',
-  styleUrl: './usuario-dados.component.scss'
+  selector: 'disciplinas',
+  templateUrl: './disciplinas.component.html',
+  styleUrl: './disciplinas.component.scss'
 })
-export class UsuarioDadosComponent implements OnInit {
+export class DisciplinasComponent implements OnInit {
   usuario!: Usuario | null;
 
   constructor(private rotaAtual: ActivatedRoute, private roteador: Router, private localStorageService: LocalStorageService, private usuarioService: UsuarioService) { }
@@ -18,4 +17,5 @@ export class UsuarioDadosComponent implements OnInit {
   ngOnInit(): void {
     this.usuario = this.localStorageService.lerUsuario();
   }
+
 }
