@@ -1,15 +1,15 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { Usuario } from '../../shared/models/usuario';
-import { LocalStorageService } from '../../shared/services/local-storage.service';
-import { UsuarioService } from '../../shared/services/usuario.service';
+import { Usuario } from '../../../shared/models/usuario';
+import { LocalStorageService } from '../../../shared/services/local-storage.service';
+import { UsuarioService } from '../../../shared/services/usuario.service';
 
 @Component({
-  selector: 'app-usuario-progresso',
-  templateUrl: './usuario-progresso.component.html',
-  styleUrl: './usuario-progresso.component.scss'
+  selector: 'progresso',
+  templateUrl: './progresso.component.html',
+  styleUrl: './progresso.component.scss'
 })
-export class UsuarioProgressoComponent implements OnInit {
+export class ProgressoComponent implements OnInit {
   usuario!: Usuario | null;
 
   constructor(private rotaAtual: ActivatedRoute, private roteador: Router, private localStorageService: LocalStorageService, private usuarioService: UsuarioService) { }
@@ -17,6 +17,5 @@ export class UsuarioProgressoComponent implements OnInit {
   ngOnInit(): void {
     this.usuario = this.localStorageService.lerUsuario();
   }
-
 
 }
