@@ -1,16 +1,15 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { Usuario } from '../../shared/models/usuario';
-import { LocalStorageService } from '../../shared/services/local-storage.service';
-import { UsuarioService } from '../../shared/services/usuario.service';
-
+import { Usuario } from '../../../shared/models/usuario';
+import { LocalStorageService } from '../../../shared/services/local-storage.service';
+import { UsuarioService } from '../../../shared/services/usuario.service';
 
 @Component({
-  selector: 'app-usuario-tarefas',
-  templateUrl: './usuario-tarefas.component.html',
-  styleUrl: './usuario-tarefas.component.scss'
+  selector: 'tarefas',
+  templateUrl: './tarefas.component.html',
+  styleUrl: './tarefas.component.scss'
 })
-export class UsuarioTarefasComponent implements OnInit {
+export class TarefasComponent implements OnInit {
   usuario!: Usuario | null;
 
   constructor(private rotaAtual: ActivatedRoute, private roteador: Router, private localStorageService: LocalStorageService, private usuarioService: UsuarioService) { }
